[feat] #61 디테일뷰에서 즐겨찾기 추가할 수 있도록 수정
