+++ app/src/main/java/com/cecd/exitmed/presentation/textSearch/SearchViewModel.kt
@@ -2,7 +2,6 @@ package com.cecd.exitmed.presentation.textSearch
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.cecd.exitmed.data.model.request.RequestBookmark
 import com.cecd.exitmed.domain.repository.BookmarkRepository
 import com.cecd.exitmed.domain.repository.TextSearchRepository
 import com.cecd.exitmed.domain.type.SearchPill
@@ -16,8 +15,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SearchViewModel @Inject constructor(
-    private val textSearchRepository: TextSearchRepository,
-    private val bookmarkRepository: BookmarkRepository
+    private val textSearchRepository: TextSearchRepository
 ) : ViewModel() {
     val searchText = MutableStateFlow("")
     private var _searchCount = MutableStateFlow<Int?>(null)
@@ -49,19 +47,6 @@ class SearchViewModel @Inject constructor(
             textSearchRepository.fetchRecentSearchTerm()
                 .onSuccess { recentSearchTerms ->
                     _recentSearchTermsState.value = UiState.Success(recentSearchTerms)
-                }
-                .onFailure { throwable ->
-                    Timber.e(throwable.message)
-                }
-        }
-    }
-
-    // TODO detail view로 이동
-    fun bookmark(pillItemSeq: Int) {
-        viewModelScope.launch {
-            bookmarkRepository.bookmark(RequestBookmark(pillItemSeq))
-                .onSuccess { isBookmarked ->
-                    Timber.tag("bookmark").e(isBookmarked.toString())
                 }
                 .onFailure { throwable ->
                     Timber.e(throwable.message)
